third_party/blink/web_tests/fast/filesystem/filesystem-missing-arguments-expected.txt

Test how file system methods react to too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS webkitRequestFileSystem() threw exception TypeError: Failed to execute 'webkitRequestFileSystem' on 'Window': 3 arguments required, but only 0 present..
PASS webkitRequestFileSystem(TEMPORARY) threw exception TypeError: Failed to execute 'webkitRequestFileSystem' on 'Window': 3 arguments required, but only 1 present..
PASS webkitResolveLocalFileSystemURL() threw exception TypeError: Failed to execute 'webkitResolveLocalFileSystemURL' on 'Window': 2 arguments required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
